QFw: factor the OpenQASM->IQM transcode into a shared util
Move the standalone circuit-transcode functions (build_iqm_circuit and its
helpers: serialize_qiskit_to_iqm, build_manual_iqm_circuit, load_qiskit_circuit,
logical_to_physical_qubits, active_qubits, load_iqm_pulse_module, parse_ref,
eval_angle, split_qasm_statements, to_jsonable) out of svc_iqm_qpm/util_iqm.py
into services/util/iqm_transcode.py, so the QRMI/QDMI shim can build IQM
circuits from OpenQASM the same way the native service does. util_iqm re-imports
the three names it still uses (to_jsonable, active_qubits, build_iqm_circuit) and
drops now-unused imports. No behavior change.
